# Introduce alternative to SkippableTokenAwareRuleWalker and scanAllTokens