--- a/src/language/utils.ts
+++ b/src/language/utils.ts
@@ -247,7 +247,8 @@ export interface TokenPosition {
     /** The end of the token */
     end: number;
 }
-export type ForEachCallback = (fullText: string, kind: ts.SyntaxKind, pos: TokenPosition) => void;
+export type ForEachTokenCallback = (fullText: string, kind: ts.SyntaxKind, pos: TokenPosition, parent: ts.Node) => void;
+export type ForEachCommentCallback = (fullText: string, kind: ts.SyntaxKind, pos: TokenPosition) => void;
 export type FilterCallback = (node: ts.Node) => boolean;
 
 /**
@@ -256,16 +257,16 @@ export type FilterCallback = (node: ts.Node) => boolean;
  * @description JsDoc comments are treated like regular comments and only visited if `skipTrivia` === false.
  * 
  * @param node The node whose tokens should be visited
- * @param skipTrivia If set to false all trivia owned by `node` or its children is included
+ * @param skipTrivia If set to false all trivia preceeding `node` or any of its children is included
  * @param cb Is called for every token of `node`. It gets the full text of the SourceFile and the position of the token within that text.
  * @param filter If provided, will be called for every Node and Token found. If it returns false `cb` will not be called for this subtree.
  */
-export function forEachToken(node: ts.Node, skipTrivia: boolean, cb: ForEachCallback, filter?: FilterCallback) {
+export function forEachToken(node: ts.Node, skipTrivia: boolean, cb: ForEachTokenCallback, filter?: FilterCallback) {
     // this function will most likely be called with SourceFile anyways, so there is no need for an additional parameter
     const sourceFile = node.getSourceFile();
     const fullText = sourceFile.text;
-    const scanner = skipTrivia ? undefined : ts.createScanner(sourceFile.languageVersion, false, sourceFile.languageVariant, fullText);
     const iterateFn = filter === undefined ? iterateChildren : iterateWithFilter;
+    const handleTrivia = skipTrivia ? undefined : createTriviaHandler(sourceFile, cb);
 
     iterateFn(node);
 
@@ -294,47 +295,113 @@ export function forEachToken(node: ts.Node, skipTrivia: boolean, cb: ForEachCall
         const tokenStart = token.getStart(sourceFile);
         if (!skipTrivia && tokenStart !== token.pos) {
             // we only have to handle trivia before each token, because there is nothing after EndOfFileToken
-            handleTrivia(token.pos, tokenStart);
+            handleTrivia!(token.pos, tokenStart, token);
         }
-        return cb(fullText, token.kind, {tokenStart, fullStart: token.pos, end: token.end});
+        return cb(fullText, token.kind, {tokenStart, fullStart: token.pos, end: token.end}, token.parent!);
     }
+}
 
+function createTriviaHandler(sourceFile: ts.SourceFile, cb: ForEachTokenCallback) {
+    const fullText = sourceFile.text;
+    const scanner = ts.createScanner(sourceFile.languageVersion, false, sourceFile.languageVariant, fullText);
     /** 
      * Scan the specified range to get all trivia tokens.
      * This includes trailing trivia of the last token and the leading trivia of the current token
      */
-    function handleTrivia(start: number, end: number) {
-        scanner!.setTextPos(start);
+    function handleTrivia(start: number, end: number, token: ts.Node) {
+        const parent = token.parent!;
+        // prevent false positives by not scanning inside JsxText
+        if (!canHaveLeadingTrivia(token.kind, parent)) {
+            return;
+        }
+        scanner.setTextPos(start);
         let position: number;
         // we only get here if start !== end, so we can scan at least one time
         do {
-            const kind = scanner!.scan();
-            position = scanner!.getTextPos();
-            cb(fullText, kind, {tokenStart: scanner!.getTokenPos(), end: position, fullStart: start});
+            const kind = scanner.scan();
+            position = scanner.getTextPos();
+            cb(fullText, kind, {tokenStart: scanner.getTokenPos(), end: position, fullStart: start}, parent);
         } while (position < end);
     }
+
+    return handleTrivia;
 }
 
-// TODO whitespaceRule template expression
-
 /** Iterate over all comments owned by `node` or its children */
-export function forEachComment(node: ts.Node, cb: ForEachCallback) {
-    // get all tokens and skip trivia. comment ranges between tokens are parsed without the need of a scanner
-    return forEachToken(node, true, (fullText, _kind, pos) => {
-        // Comments before the first token (pos.fullStart === 0) are all considered leading comments, so no need for special treatment
-        ts.forEachLeadingCommentRange(fullText, pos.fullStart, commentCallback);
-        ts.forEachTrailingCommentRange(fullText, pos.end, commentCallback);
+export function forEachComment(node: ts.Node, cb: ForEachCommentCallback) {
+    /* Visit all tokens and skip trivia.
+       Comment ranges between tokens are parsed without the need of a scanner.
+       forEachToken also does intentionally not pay attention to the correct comment ownership of nodes as it always
+       scans all trivia before each token, which could include trailing comments of the previous token.
+       Comment onwership is done right in this function*/
+    return forEachToken(node, true, (fullText, tokenKind, pos, parent) => {
+        // don't search for comments inside JsxText
+        if (canHaveLeadingTrivia(tokenKind, parent)) {
+            // Comments before the first token (pos.fullStart === 0) are all considered leading comments, so no need for special treatment
+            ts.forEachLeadingCommentRange(fullText, pos.fullStart, commentCallback);
+        }
+        if (canHaveTrailingTrivia(tokenKind, parent)) {
+            ts.forEachTrailingCommentRange(fullText, pos.end, commentCallback);
+        }
         function commentCallback(tokenStart: number, end: number, kind: ts.SyntaxKind) {
             cb(fullText, kind, {tokenStart, end, fullStart: pos.fullStart});
         }
     });
 }
 
+/** Exclude leading positions that would lead to scanning for trivia inside JsxText */
+function canHaveLeadingTrivia(tokenKind: ts.SyntaxKind, parent: ts.Node): boolean {
+    if (tokenKind === ts.SyntaxKind.JsxText) {
+        return false; // there is no trivia before JsxText
+    }
+    if (tokenKind === ts.SyntaxKind.OpenBraceToken) {
+        // before a JsxExpression inside a JsxElement's body can only be other JsxChild, but no trivia
+        return parent.kind !== ts.SyntaxKind.JsxExpression || parent.parent!.kind !== ts.SyntaxKind.JsxElement;
+    }
+    if (tokenKind === ts.SyntaxKind.LessThanToken) {
+        if (parent.kind === ts.SyntaxKind.JsxClosingElement) {
+            return false; // would be inside the element body
+        }
+        if (parent.kind === ts.SyntaxKind.JsxOpeningElement || parent.kind === ts.SyntaxKind.JsxSelfClosingElement) {
+            // there can only be leading trivia if we are at the end of the top level element
+            return parent.parent!.parent!.kind !== ts.SyntaxKind.JsxElement;
+        }
+    }
+    return true;
+}
+
+/** Exclude trailing positions that would lead to scanning for trivia inside JsxText */
+function canHaveTrailingTrivia(tokenKind: ts.SyntaxKind, parent: ts.Node): boolean {
+    if (tokenKind === ts.SyntaxKind.JsxText) {
+        return false; // there is no trivia after JsxText
+    }
+    if (tokenKind === ts.SyntaxKind.CloseBraceToken) {
+        // after a JsxExpression inside a JsxElement's body can only be other JsxChild, but no trivia
+        return parent.kind !== ts.SyntaxKind.JsxExpression || parent.parent!.kind !== ts.SyntaxKind.JsxElement;
+    }
+    if (tokenKind === ts.SyntaxKind.GreaterThanToken) {
+        if (parent.kind === ts.SyntaxKind.JsxOpeningElement) {
+            return false; // would be inside the element
+        }
+        if (parent.kind === ts.SyntaxKind.JsxClosingElement || parent.kind === ts.SyntaxKind.JsxSelfClosingElement) {
+            // there can only be trailing trivia if we are at the end of the top level element
+            return parent.parent!.parent!.kind !== ts.SyntaxKind.JsxElement;
+        }
+    }
+    return true;
+}
+
 function hasCommentCallback() {
     return true;
 }
 
-/** Checks if there are any comments between `position` and the next non-tivia token */
+/** 
+ * Checks if there are any comments between `position` and the next non-tivia token
+ * 
+ * @param text The text to scan
+ * @param position The position inside `text` where to start scanning. Make sure that this is a valid start position.
+ *                 This value is typically obtained from `node.getFullStart()` or `node.getEnd()`
+ */
 export function hasCommentAfterPosition(text: string, position: number): boolean {
     return ts.forEachTrailingCommentRange(text, position, hasCommentCallback) ||
            ts.forEachLeadingCommentRange(text, position, hasCommentCallback) ||