--- a/src/rules/whitespaceRule.ts
+++ b/src/rules/whitespaceRule.ts
@@ -68,7 +68,7 @@ export class Rule extends Lint.Rules.AbstractRule {
     }
 }
 
-class WhitespaceWalker extends Lint.SkippableTokenAwareRuleWalker {
+class WhitespaceWalker extends Lint.RuleWalker {
     private scanner: ts.Scanner;
 
     constructor(sourceFile: ts.SourceFile, options: Lint.IOptions) {
@@ -80,27 +80,14 @@ class WhitespaceWalker extends Lint.SkippableTokenAwareRuleWalker {
         super.visitSourceFile(node);
 
         let prevTokenShouldBeFollowedByWhitespace = false;
-        this.scanner.setTextPos(0);
-
-        Lint.scanAllTokens(this.scanner, (scanner: ts.Scanner) => {
-            const startPos = scanner.getStartPos();
-            const tokenKind = scanner.getToken();
-
+        Lint.forEachToken(node, false, (_text, tokenKind, pos) => {
             if (tokenKind === ts.SyntaxKind.WhitespaceTrivia || tokenKind === ts.SyntaxKind.NewLineTrivia) {
                 prevTokenShouldBeFollowedByWhitespace = false;
+                return;
             } else if (prevTokenShouldBeFollowedByWhitespace) {
-                this.addMissingWhitespaceErrorAt(startPos);
+                this.addMissingWhitespaceErrorAt(pos.tokenStart);
                 prevTokenShouldBeFollowedByWhitespace = false;
             }
-
-            const skip = this.getSkipEndFromStart(startPos);
-            if (skip !== undefined) {
-                // tokens to skip are places where the scanner gets confused about what the token is, without the proper context
-                // (specifically, regex, identifiers, and templates). So skip those tokens.
-                scanner.setTextPos(skip);
-                return;
-            }
-
             // check for trailing space after the given tokens
             switch (tokenKind) {
                 case ts.SyntaxKind.CatchKeyword:
@@ -139,6 +126,10 @@ class WhitespaceWalker extends Lint.SkippableTokenAwareRuleWalker {
                 default:
                     break;
             }
+        },
+        (n) => {
+            // don't apply the whitespace rules to JSX specific syntax
+            return n.kind !== ts.SyntaxKind.JsxElement && n.kind !== ts.SyntaxKind.JsxSelfClosingElement;
         });
     }
 
@@ -217,16 +208,6 @@ class WhitespaceWalker extends Lint.SkippableTokenAwareRuleWalker {
             this.checkForTrailingWhitespace(position);
         }
         super.visitImportEqualsDeclaration(node);
-    }
-
-    public visitJsxElement(node: ts.JsxElement) {
-        this.addTokenToSkipFromNode(node);
-        super.visitJsxElement(node);
-    }
-
-    public visitJsxSelfClosingElement(node: ts.JsxSelfClosingElement) {
-        this.addTokenToSkipFromNode(node);
-        super.visitJsxSelfClosingElement(node);
     }
 
     public visitTypeAssertionExpression(node: ts.TypeAssertion) {