… test/rules/_integration/react/test.tsx.lint
… test/rules/comment-format/lower/test.tsx.lint
… test/rules/whitespace/all/test.ts.fix
… test/rules/whitespace/all/test.ts.lint