#### PR checklist

- [x] Addresses an existing issue: #2007
- [x] New feature, bugfix, or enhancement
  - [x] Includes tests
- [ ] Documentation update
- [x] Enable CircleCI for your fork (https://circleci.com/add-projects)

#### What changes did you make?

Scanning the whole source file is very fragile, because the grammar is not context free. You had to skip certain ranges, that were known to cause trouble. For example regex literals or template expressions.
This approach is prone to errors, as it didn't cover all cases and has to adopt new grammar as the language adds it.

The new function hands the scanning and parsing off to the parser and only iterates over the tokens of all AST nodes. Skipping ranges is no longer necessary.

This fixes #2007 and replaces #2009 (I took 2 commits from that PR to include the contained tests)

For more information please refer to the description of each individual commit.

#### Is there anything you'd like reviewers to focus on?

This is a rather big PR. I could split it into several smaller ones if you want.

`WhitespaceRule` used to skip `TemplateExpression`, `JsxElement` and `JsxSelfClosingElement`. That's no longer necessary.
That means not all whitespace errors were found with the previous implementation. Now we find all errors, even inside `JsxExpression` and `TemplateExpression`.
If you think this is a breaking change, I would fall back to skipping the aforementioned nodes for now and submit a separate PR to include them for the next major version.


I'm currently investigating what the performance impact of this change is. I'll keep you posted.