@ajafff 2017-01-14
### Performance

Linting `tslint`'s code with itself, repeated several times.

master: 10.9 sec to 11.4 sec
this branch: 9.7 sec to 10.0 sec

Seems like a 10% overall speedup as a side effect of a bugfix isn't that bad, is it?

### Where's that coming from?

* We no longer need `SkippableTokenAwareRuleWalker`. See #2006 for some perf numbers
* `node.createChildren` only needs to rescan "synthetic tokens" - tokens that are not included in the AST - instead of scanning the whole SourceFile.
* `node.getChildren` only calls `node.createChildren` once and then reuses this array. So `EnableDisableRuleWalker` does the "heavy lifting" of creating those arrays for the whole AST. As long as the AST is not recreated (e.g. because a fix was applied), all rules using `node.getChildren` (and similar methods like `getChildAt`, `getFirstToken`, ...) benefit from the cached child array.