@nchen63 2017-01-16
as far as the new errors it finds, I'd consider those as fixes which aren't breaking. @adidahiya for comment